instance variable name is chef_version, not require_chef_version
`UNSET_VALUE` makes it possible to know the difference between a 
configuration value that was never set, and a value that is nil 
(explicitly). A downstream middleware uses the value in `chef_version`
to determine if the plugin should operate in no-op mode; if the value is
`nil` the plugin silently skips.

This fixes issue #3.