improve package install logic. fixes #33
Splitting the fetching of `install.sh` from the actual execution of
said script will improve the user experience. It will also allow us to
modify the `install.sh` if needed (see #22).

We also switched to using Vagrant's built in `Vagrant::Util::Downloader`
class which is a pure-Ruby downloader which removes requirement of the
guest OS having `wget` or `curl` installed.