Change :latest to not hit rubygems
Just send the string "latest" into the omnibus API.  This makes it
not idempotent so that it will run install.sh every time.  Since we
hit rubygems before we're no better off in disconnected mode that
we were before.

Tested on Ubuntu and CentOS, the acceptance tests for windows are
trolling me hard and I don't know how to fix.